Fix the MD5 code for 64-bit
http://sf.net/tracker/?func=detail&aid=2789622&group_id=22615&atid=375631

+++ src/md5.h
@@ -22,14 +22,16 @@ These notices must be retained in any copies of any part of this
 documentation and/or software.
  */
 
+#include <stdint.h>
+
 /* POINTER defines a generic pointer type */
 typedef unsigned char *POINTER;
 
 /* UINT2 defines a two byte word */
-typedef unsigned short int UINT2;
+typedef uint16_t UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* MD5 context. */
 typedef struct {
